refactor: Restructure code
