+++ src/pages/about/api.tsx
@@ -2,7 +2,7 @@ import { Code, Heading, Link, Text, VStack } from "@chakra-ui/react";
 import { withTheme } from "@emotion/react";
 import Layout from "components/Layout";
 import { EXAMPLES } from "lib/examples";
-import splitParts from "lib/split-parts";
+import splitParts from "lib/utils/split-parts";
 import libnpmdiff from "libnpmdiff";
 import { GetStaticProps, NextPage } from "next";
