+++ src/pages/api/[...parts].ts
@@ -1,9 +1,9 @@
 import destination from "lib/destination";
-import parseQuery from "lib/parse-query";
-import rawQuery from "lib/raw-query";
-import setCacheControl from "lib/set-cache-control";
-import specsToDiff from "lib/specs-to-diff";
-import splitParts from "lib/split-parts";
+import parseQuery from "lib/utils/parse-query";
+import rawQuery from "lib/utils/raw-query";
+import setCacheControl from "lib/utils/set-cache-control";
+import specsToDiff from "lib/utils/specs-to-diff";
+import splitParts from "lib/utils/split-parts";
 import libnpmdiff from "libnpmdiff";
 import { NextApiHandler } from "next";
